refactor: split LetterboxdClient into HttpClient, Auth, Scraper, and Diary classes
